Remove dynamic build
Dynamic build was set to enhance compilation time.
The only issue is that it made running `cargo build --release`
created binaries that required `bevy_dylib.so` which is not ideal.

By simply removing it, the binary doesn't depend on a dynamic library.
And we can still benefit from dynamic builds by using `cargo run
--features bevy/dynamic`