Restrict git revision walk to a single commit
We were going through the complete list of commits.
This was unnecessary for the current application design
and was slowing down the extraction process considerably.

Maybe when we decide to extract data from every commits we will
perform a parallel revwalk again. It is also important to note
that we have spent a lot of time trying to figure out how we could
parallelize the revwalk process but it seemed to be very difficult
because the git2::Repository struc does not implement the Send and Sync
traits.

We are glad this solution was found as it speed up the extraction
process